requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
django-ninja
psycopg2-binary
python-dotenv
gunicorn

pytest
pytest-django
djangorestframework
pytest-cov
